--- tutorials/BrowserView.wrapSync.md
@@ -1,6 +1,0 @@
-Synchronously returns a BrowserView object that represents an existing application.
-# Example
-```js
-const browserView = fin.BrowserView.wrapSync({ uuid: 'testBrowserView', name: 'testBrowserViewName' });
-await browserView.hide();
-```
