+++ tutorials/System.EventEmitter.md
@@ -111,11 +111,11 @@ fin.System.removeAllListeners("monitor-info-changed");
 * window-user-movement-enabled (see {@tutorial Window.EventEmitter})
 * window-will-move (see {@tutorial Window.EventEmitter})
 * window-will-resize (see {@tutorial Window.EventEmitter})
-* view-created (see {@tutorial BrowserView.EventEmitter})
-* view-attached (see {@tutorial BrowserView.EventEmitter})
-* view-shown (see {@tutorial BrowserView.EventEmitter})
-* view-hidden (see {@tutorial BrowserView.EventEmitter})
-* view-destroyed (see {@tutorial BrowserView.EventEmitter})
+* view-created (see {@tutorial View.EventEmitter})
+* view-attached (see {@tutorial View.EventEmitter})
+* view-shown (see {@tutorial View.EventEmitter})
+* view-hidden (see {@tutorial View.EventEmitter})
+* view-destroyed (see {@tutorial View.EventEmitter})
 
 ### System Events
