+++ tutorials/View.wrapSync.md
@@ -1,0 +1,6 @@
+Synchronously returns a View object that represents an existing application.
+# Example
+```js
+const browserView = fin.View.wrapSync({ uuid: 'testView', name: 'testViewName' });
+await browserView.hide();
+```
